fix: render meta title as OGP
